Extract more business logic into dedicated services
